Example Executable
This is a concept of an example-hosting executable. It includes some static variable usage to help get a feel for how something like this could be structured.

I really wanted to construct a `Package.swift` that could build the executable with three possible configurations:

- Swift 5 compiler (5 mode)
- Swift 6 compiler (5 mode)
- Swift 6 compiler (6 mode)

This turned out to be pretty tricky to pull off, but I have it working. SPM will not let you use the same path for multiple targets, but I also was able to avoid duplicating the code by using symlinks.